update: fix code and add more test for size food

+++ coffee_shop_management_backend/module/sizefood/sizefoodmodel/size_food.go
@@ -10,8 +10,8 @@ type SizeFood struct {
 	FoodId   string             `json:"foodId" gorm:"column:foodId;"`
 	SizeId   string             `json:"sizeId" gorm:"column:sizeId;"`
 	Name     string             `json:"name" gorm:"column:name;"`
-	Cost     float32            `json:"cost" gorm:"column:cost;"`
-	Price    float32            `json:"price" gorm:"column:price;"`
+	Cost     int                `json:"cost" gorm:"column:cost;"`
+	Price    int                `json:"price" gorm:"column:price;"`
 	RecipeId string             `json:"-" gorm:"column:recipeId;"`
 	Recipe   recipemodel.Recipe `json:"recipe" gorm:"foreignkey:recipeId"`
 }
