+++ coffee_shop_management_backend/module/sizefood/sizefoodmodel/size_food_create.go
@@ -9,8 +9,8 @@ type SizeFoodCreate struct {
 	FoodId   string                    `json:"-" gorm:"column:foodId;"`
 	SizeId   string                    `json:"-" gorm:"column:sizeId;"`
 	Name     string                    `json:"name" gorm:"column:name"`
-	Cost     float32                   `json:"cost" gorm:"column:cost"`
-	Price    float32                   `json:"price" gorm:"column:price"`
+	Cost     int                       `json:"cost" gorm:"column:cost"`
+	Price    int                       `json:"price" gorm:"column:price"`
 	RecipeId string                    `json:"-" gorm:"column:recipeId;"`
 	Recipe   *recipemodel.RecipeCreate `json:"recipe" gorm:"-"`
 }
